Plot survey index list (e.g. retrospective analysis)

Plot survey index list (e.g. retrospective analysis)

plot_simulation_list(
  x,
  base = 1,
  rescale = FALSE,
  lwd = 1.5,
  main = NULL,
  allCI = FALSE,
  includeCI = TRUE,
  ylim = NULL
)

Arguments

x

(named) list of "surveyIdx" objects for example from "retro.surveyIdx" or "leaveout.surveyIdx"

base

Either index of x that should considered the "base run" (integer), OR object of class "surveyIdx". Confidence bounds will be shown for this model only.

rescale

Should indices be rescaled to have mean 1 (over the set of intersecting years)? Default: FALSE

lwd

line width argument to plot

main

if not NULL override main plotting default title of "Age group a"

allCI

show 95% confidence lines for all indices? Default FALSE.

includeCI

Show confidence intervals? Default TRUE.

ylim

Y axis range. If NULL (default) then determine automatically.

Value

nothing